Add files via upload
This is a online shopping website which helps the customers to compare the prices of a product with different stores available in that area. The intension of creating this website is to provide best price of a product to the customers and save the time of customers to compare the prices by going through different stores.